% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questions.R
\name{combine_questions}
\alias{combine_questions}
\title{Join questions and follow-ups}
\usage{
combine_questions(q)
}
\arguments{
\item{q}{A data frame with questions.}
}
\description{
Join questions and follow-ups
}
\examples{
q <-  get_sdm_from_env("default_questions")
q$followup_level[5] <- 3
combine_questions(q)

}
