% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{components}
\alias{components}
\title{Components}
\format{
A data frame with component configuration (some columns are lists).
}
\usage{
components
}
\description{
Components
}
\keyword{datasets}
