% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_connect}
\alias{db_connect}
\title{Connect to Database}
\usage{
db_connect(...)
}
\arguments{
\item{...}{Arguments bassed to \code{\link[DBI:dbConnect]{DBI::dbConnect()}}.}
}
\value{
A database connection.
}
\description{
Connect to Database
}
