% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_deployment_materials}
\alias{db_deployment_materials}
\title{Get Deployment Materials from DB}
\usage{
db_deployment_materials(con, modelid, deploymentid)
}
\arguments{
\item{con}{DB connection.}

\item{modelid}{Model ID.}

\item{deploymentid}{Deployment ID.}
}
\value{
A data frame with joined deployment materials.
}
\description{
Get Deployment Materials from DB
}
