% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_disconnect}
\alias{db_disconnect}
\title{Disconnect from Database}
\usage{
db_disconnect(con, ...)
}
\arguments{
\item{con}{DB connection.}

\item{...}{Arguments passed to \code{\link[DBI:dbDisconnect]{DBI::dbDisconnect()}}.}
}
\description{
Disconnect from Database
}
