% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_read_models}
\alias{db_read_models}
\title{Get models from DB}
\usage{
db_read_models(con, model_id = NULL)
}
\arguments{
\item{con}{DB connection.}

\item{model_id}{Model ID (optional).}
}
\value{
A data frame with models. Optionally filtered to model_id.
}
\description{
Get models from DB
}
