% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_read_species}
\alias{db_read_species}
\title{Get species from DB}
\usage{
db_read_species(con, species_id = NULL, lang = "english")
}
\arguments{
\item{con}{DB connection.}

\item{species_id}{Species ID (optional).}
}
\value{
A data frame with species and display names. Optionally filtered to
species_id.
}
\description{
Get species from DB
}
