% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_read_table}
\alias{db_read_table}
\title{Get Table from DB}
\usage{
db_read_table(con, table_name)
}
\arguments{
\item{con}{DB connection.}

\item{table_name}{Table name.}
}
\value{
A data frame with the table data.
}
\description{
Get Table from DB
}
