% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_read_user_info}
\alias{db_read_user_info}
\title{Get User Info from DB}
\usage{
db_read_user_info(con, userid, deploymentid)
}
\arguments{
\item{con}{DB connection.}

\item{userid}{User ID.}

\item{deploymentid}{Deployment ID.}
}
\value{
A data frame with user info and access roles.
The \code{"user_roles"} attribute gives the user roles as a vector.
}
\description{
Get User Info from DB
}
