% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_user_info}
\alias{db_user_info}
\title{Get User Info from DB}
\usage{
db_user_info(con, userid, deploymentid)
}
\arguments{
\item{con}{DB connection.}

\item{userid}{User ID.}

\item{deploymentid}{Deployment ID.}
}
\value{
A data frame with user info and access roles.
}
\description{
Get User Info from DB
}
