% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{default_questions}
\alias{default_questions}
\title{Default Questions}
\format{
A data frame with default questions.
}
\usage{
default_questions
}
\description{
Default Questions
}
\keyword{datasets}
