% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{default_species_table_canada}
\alias{default_species_table_canada}
\title{Default Species Table}
\usage{
default_species_table_canada()
}
\value{
Data frame.
}
\description{
Default Species Table
}
\examples{
str(default_species_table_canada())

}
