% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake.R
\name{fake}
\alias{fake}
\alias{fake_comment}
\alias{fake_evaluation}
\alias{fake_response}
\title{Fake DB Entries}
\usage{
fake_comment(deployment_id, model_id, species_id, user_id)

fake_evaluation(
  deployment_id,
  material_id,
  component_id,
  user_id,
  questions = NULL
)

fake_response(question)
}
\arguments{
\item{deployment_id}{Deployment ID.}

\item{model_id}{Model ID.}

\item{species_id}{Species ID.}

\item{user_id}{User ID.}

\item{material_id}{Material ID.}

\item{component_id}{Component ID.}

\item{questions}{A questions data frame.}

\item{question}{A questions data frame with a single row.}
}
\description{
Fake DB Entries
}
\examples{
fake_comment(
    deployment_id = "deployment1",
    model_id = "model1",
    species_id = "MOTR",
    user_id = "draper")

str(fake_response(sdmEvalToolCore::default_questions[1,]))

fake_evaluation(
    deployment_id = "deployment1",
    material_id = "material1",
    component_id = "model_fit",
    user_id = "draper")

}
