% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{followup_questions}
\alias{followup_questions}
\title{Follow-up Questions}
\format{
A data frame with follow-up questions.
}
\usage{
followup_questions
}
\description{
Follow-up Questions
}
\keyword{datasets}
