% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_auth.R
\name{generate_password}
\alias{generate_password}
\title{Generate password}
\usage{
generate_password(length = 12, start_with_letter = TRUE)
}
\arguments{
\item{length}{Length of the password. Default is 12.}

\item{start_with_letter}{Logical indicating if the password should start
with a letter. Default is TRUE.}
}
\value{
A randomly generated password.
}
\description{
Generate password
}
\examples{
generate_password()

}
