% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{get_comp_ready}
\alias{get_comp_ready}
\title{Which Components Are Ready}
\usage{
get_comp_ready(ready = character(0L))
}
\arguments{
\item{ready}{Character, component names that are ready.}
}
\value{
A data frame with components and a column telling which ones are ready.
The \code{"percent_ready"} attribute gives a percentage value of the
mandatory components that are ready.
}
\description{
Which Components Are Ready
}
\examples{
r <- get_comp_ready(c("observations", "model_metadata"))
r
attr(r, "percent_ready")

}
