% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{get_comp_rule}
\alias{get_comp_rule}
\title{Get Rule for a Component}
\usage{
get_comp_rule(
  component_id,
  rule_type = c("upload", "display", "evaluation", "reporting")
)
}
\arguments{
\item{component_id}{Character, component name (length 1).}

\item{rule_type}{CHaracter, the type of rule.}
}
\value{
A list with the rules, or \code{NULL}.
}
\description{
Get Rule for a Component
}
\examples{
str(get_comp_rule("observations", "upload"))
str(get_comp_rule("observations", "reporting"))

}
