% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_fields}
\alias{get_fields}
\title{Get Table Fields}
\usage{
get_fields(table_name)
}
\arguments{
\item{table_name}{Table name.}
}
\value{
A data frame.
}
\description{
Get Table Fields
}
\examples{
get_fields("users")
get_fields("components")
}
