% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{get_file_ext}
\alias{get_file_ext}
\title{Get File Extension}
\usage{
get_file_ext(path)
}
\arguments{
\item{path}{Character, a file path.}
}
\value{
Lower case file extension as string.
}
\description{
Get File Extension
}
\examples{
get_file_ext("bam_v5/oven/observations.parquet")

}
