% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{get_sdm_from_env}
\alias{get_sdm_from_env}
\title{Get default or custom configuration}
\usage{
get_sdm_from_env(component_name)
}
\value{
A data frame with default components
}
\description{
Get default or custom configuration
}
