% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{get_user_roles}
\alias{get_user_roles}
\title{Get User Roles}
\usage{
get_user_roles(role)
}
\arguments{
\item{role}{Character, 1 or multiple user roles.}
}
\value{
A data frame with the permissible user actions (TRUE/FALSE).
}
\description{
Get User Roles
}
\examples{
str(get_user_roles(c("modeler")))
str(get_user_roles(c("evaluator")))
str(get_user_roles(c("evaluator", "modeler")))

}
