% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{make_dir}
\alias{make_dir}
\title{Make Directory}
\usage{
make_dir(path, ...)
}
\arguments{
\item{path}{Character, a file path.}

\item{...}{Arguments passed to \code{\link[=dir.create]{dir.create()}}.}
}
\value{
Create a directory recursively as a side effect,
returning logical indicating the result ().
}
\description{
Make Directory
}
