% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{make_material_id}
\alias{make_material_id}
\title{Make Material ID}
\usage{
make_material_id(model_id, species_id, component_id, sep = "_")
}
\arguments{
\item{model_id}{Model ID.}

\item{species_id}{Species ID.}

\item{component_id}{Component ID.}

\item{sep}{Character. Separator.}
}
\description{
Make Material ID
}
\examples{
make_material_id("model1", NA, "predictor_metadata")
make_material_id("model1", "CAWA", "predictor_raster")

}
