% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{make_target_path}
\alias{make_target_path}
\title{Make Target Path}
\usage{
make_target_path(path, data = list(), base = NULL)
}
\arguments{
\item{path}{Character, a file path.}

\item{data}{List with name value pairs, the values are substituted
into the path according to glue rules (with curly braces).}

\item{base}{Character, base path, defaults to \code{sdmevaltool_options()$base}
when \code{NULL}.}
}
\value{
The path string with data values substituted.
}
\description{
Make Target Path
}
\examples{
make_target_path("bam_v5/oven/observations.parquet")
make_target_path("{model}/{species}/observations.parquet",
  list(model = "bam_v5", species = "oven"))
make_target_path("{model}/{species}/observations.parquet",
  list(model = "bam_v5", species = "oven"), ".")

}
