% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_auth.R
\name{make_user_db}
\alias{make_user_db}
\title{Make user database suitable for shinymanager}
\usage{
make_user_db()
}
\value{
A data frame with user credentials.
}
\description{
Make user database suitable for shinymanager
}
