% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{prepare_material_entry}
\alias{prepare_material_entry}
\title{Prepare Material Entry}
\usage{
prepare_material_entry(
  model_id,
  species_id,
  component_id,
  user_id,
  material_settings = "[]"
)
}
\arguments{
\item{model_id}{Model ID.}

\item{species_id}{Species ID.}

\item{component_id}{Component ID.}

\item{user_id}{User ID.}

\item{material_settings}{Material settings.}
}
\description{
Prepare Material Entry
}
\examples{
prepare_material_entry("model1", "CAWA", "predictor_raster", "testuser")

}
