% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{read_file}
\alias{read_file}
\title{Read File}
\usage{
read_file(path, ...)
}
\arguments{
\item{path}{Character, a file path.}

\item{...}{Arguments passed to the reader functions.}
}
\value{
Return the result from the file.
}
\description{
Read File
}
