% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{sdmevaltool_options}
\alias{sdmevaltool_options}
\title{Package Options}
\usage{
sdmevaltool_options(...)
}
\arguments{
\item{...}{Names list of options.}
}
\description{
Package Options
}
\examples{
(sdmevaltool_options())

}
