% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materials.R
\name{upload-deployments}
\alias{upload-deployments}
\alias{prep_deployment_questions}
\alias{prep_deployment_settings}
\alias{prep_deployment_subunits}
\title{Prepare Deployment Materials}
\usage{
prep_deployment_questions(deployment_id, x = NULL, ...)

prep_deployment_settings(deployment_id, x, ...)

prep_deployment_subunits(
  deployment_id,
  x = NULL,
  reference_raster,
  reference_points,
  n = c(20, 20),
  square = TRUE,
  ...
)
}
\arguments{
\item{deployment_id}{Deployment ID.}

\item{x}{An object to upload.}

\item{...}{Arguments passed to the write function.}

\item{reference_raster}{Reference raster layer.}

\item{reference_points}{Reference spatial points object.}

\item{n}{Number of grid cells in x and y direction.}

\item{square}{Logical. If `FALSE``, create hexagonal grid.}
}
\description{
Prepare Deployment Materials
}
