% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{users}
\alias{users}
\title{User Roles}
\usage{
users(role)
}
\arguments{
\item{role}{Character, 1 or multiple user roles.}
}
\value{
A data frame with the permissible user actions (TRUE/FALSE).
}
\description{
User Roles
}
\examples{
str(users(c("modeler")))
str(users(c("evaluator")))
str(users(c("evaluator", "modeler")))

}
