% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{write_file}
\alias{write_file}
\title{Write File}
\usage{
write_file(x, path, ...)
}
\arguments{
\item{x}{Object to write.}

\item{path}{Character, a file path.}

\item{...}{Arguments passed to the writer functions.}
}
\value{
Writes file as a side effect.
}
\description{
Write File
}
