% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{check_table}
\alias{check_table}
\title{Check Table}
\usage{
check_table(x, table_name, dryrun = FALSE)
}
\arguments{
\item{x}{A data frame.}

\item{table_name}{Table name.}

\item{dryrun}{Logical, if checks fail and error is produced,
unless it is a dry run.}
}
\value{
Invisible \code{TRUE} if all check passed.
}
\description{
Check Table
}
\examples{
users <- data.frame(
    user_id = c("holden", "draper", "okoye"),
    user_name = c("James Holden", "Bobbie Draper", "Elvi Okoye"),
    user_email = c(
        "jim@rocinante.org",
        "bdraper@mcrn.gov",
        "okoye@rce.com"
    ),
    user_affiliation = c("Rocinante", "MCRN", "RCE"),
    admin = c(TRUE, FALSE, FALSE)
)
check_table(users, "users")

models <- data.frame(
    model_id = "bam_v5_can71",
    model_name = "BAM v5 Can 71",
    model_description = "BAM version 5 Canada model in BCR 71"
)
check_table(models, "models")

check_table(models, "users", dryrun = TRUE)

}
