% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_create_tables}
\alias{db_create_tables}
\title{Create DB tables}
\usage{
db_create_tables(con, tables = NULL, force = FALSE)
}
\arguments{
\item{con}{DB connection.}

\item{tables}{Character, table name.}

\item{force}{Logical, force (create even if not exists).}
}
\description{
Create DB tables
}
\examples{
con <- db_connect(":memory:")
db_create_tables(con, "models")
db_create_tables(con)
db_disconnect(con)

}
