% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_read_deployment_materials}
\alias{db_read_deployment_materials}
\title{Get Deployment Materials from DB}
\usage{
db_read_deployment_materials(con, deployment_id)
}
\arguments{
\item{con}{DB connection.}

\item{deployment_id}{Deployment ID.}
}
\value{
A data frame with joined deployment materials.
}
\description{
Get Deployment Materials from DB
}
