% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_read_evaluations}
\alias{db_read_evaluations}
\title{Get Evaluations from DB}
\usage{
db_read_evaluations(con, deployment_id = NULL, user_id = NULL)
}
\arguments{
\item{con}{DB connection.}

\item{deployment_id}{Deployment ID.}

\item{user_id}{Evaluation create user IDs (multiple allowed).}
}
\value{
A data frame with evaluations for the deployment.
}
\description{
Get Evaluations from DB
}
