% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_read_evaluations}
\alias{db_read_evaluations}
\title{Get Comments from DB}
\usage{
db_read_evaluations(con, deployment_id)
}
\arguments{
\item{con}{DB connection.}

\item{deployment_id}{Deployment ID.}
}
\value{
A data frame with evaluations for the deployment.
}
\description{
Get Comments from DB
}
