% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_write_table}
\alias{db_write_table}
\title{Write data to a table}
\usage{
db_write_table(con, table, data, validate = TRUE, verbose = TRUE)
}
\arguments{
\item{con}{A database connection.}

\item{table}{Table name.}

\item{data}{Data frame to append to the DB table.}

\item{validate}{Logical, should \code{data} be validated?}

\item{verbose}{Logical.}
}
\value{
Invisible \code{TRUE}.
}
\description{
Data will be appended to an existing DB table.
}
