% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_write_table}
\alias{db_write_table}
\title{Write Data to an Existing Table}
\usage{
db_write_table(
  con,
  table,
  data,
  mode = c("insert", "update", "upsert"),
  check = TRUE,
  dryrun = FALSE
)
}
\arguments{
\item{con}{A database connection.}

\item{table}{Table name.}

\item{data}{Data frame to append to the DB table.}

\item{mode}{How to execute the write operation:
insert (append) new record,
update existing record, or
upsert (update existing or insert if not).}

\item{check}{Logical, should \code{data} be validated?}

\item{dryrun}{Logical, write to a text file when \code{TRUE}
and to the database when \code{FALSE}.}
}
\value{
Invisible \code{TRUE}.
}
\description{
Data will be inserted/updated/upserted in an existing DB table.
}
\examples{
con <- db_connect(":memory:")
db_create_tables(con, "models")
models <- data.frame(
    model_id = "bam_v5_can71",
    model_name = "BAM v5 Can 71",
    model_description = "BAM version 5 Canada model in BCR 71"
)
db_write_table(con, "models", models, mode = "insert")
db_disconnect(con)

}
