% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_write_table}
\alias{db_write_table}
\title{Write data to a table}
\usage{
db_write_table(
  con,
  table,
  data,
  validate = TRUE,
  verbose = TRUE,
  dryrun = FALSE
)
}
\arguments{
\item{con}{A database connection.}

\item{table}{Table name.}

\item{data}{Data frame to append to the DB table.}

\item{validate}{Logical, should \code{data} be validated?}

\item{verbose}{Logical.}

\item{dryrun}{Logical.}
}
\value{
Invisible \code{TRUE}.
}
\description{
Data will be appended to an existing DB table.
}
\examples{
con <- db_connect(":memory:")
db_create_tables(con, "models", force = FALSE)
models <- data.frame(
    model_id = "bam_v5_can71",
    model_name = "BAM v5 Can 71",
    model_description = "BAM version 5 Canada model in BCR 71"
)
db_write_table(con, "models", models)
db_disconnect(con)

}
