% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{scaffold_table}
\alias{scaffold_table}
\title{Scaffold Table}
\usage{
scaffold_table(table_name)
}
\arguments{
\item{table_name}{Table name.}
}
\value{
A data frame returned invisibly, called for the printed side effects.
}
\description{
Scaffold Table
}
\examples{
scaffold_table("users")
scaffold_table("materials")

}
