% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdmEvalToolCore-package.R
\docType{package}
\name{sdmEvalToolCore-package}
\alias{sdmEvalToolCore}
\alias{sdmEvalToolCore-package}
\title{sdmEvalToolCore: SDM Evaluation Tool Core Functions}
\description{
The goal of sdmEvalToolCore is to implement the data model and common non-interactive functionality for the SDM Evaluation Tool packages.
}
\author{
\strong{Maintainer}: Peter Solymos \email{psolymos@gmail.com} (\href{https://orcid.org/0000-0001-7337-1740}{ORCID})

Authors:
\itemize{
  \item Steffi LaZerte \email{sel@steffilazerte.ca} (\href{https://orcid.org/0000-0002-7690-8360}{ORCID})
}

Other contributors:
\itemize{
  \item Analythium Solutions Inc. [copyright holder]
  \item His Majesty the King in Right of Canada as represented by the Minister of the Environment (Copyright holder for default evaluation/drill down questions.) [copyright holder]
  \item Biodiversity Pathways (Copyright holder for glossary.csv, and default evaluation/drill down questions.) [copyright holder]
  \item Alberta Biodiversity Monitoring Institute (Copyright holder for glossary.csv, and default evaluation/drill down questions.) [copyright holder]
}

}
\keyword{internal}
