% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdmEvalToolCore-package.R
\docType{package}
\name{sdmEvalToolCore-package}
\alias{sdmEvalToolCore}
\alias{sdmEvalToolCore-package}
\title{sdmEvalToolCore: SDM Evaluation Tool Core Functions}
\description{
The goal of sdmEvalToolCore is to implement the data model and non-interactive functionality for sdmEvalToolUI according to the SDM Evaluation Tool requirements.
}
\author{
\strong{Maintainer}: Peter Solymos \email{psolymos@gmail.com} (\href{https://orcid.org/0000-0001-7337-1740}{ORCID})

Authors:
\itemize{
  \item Steffi LaZerte \email{sel@steffilazerte.ca} (\href{https://orcid.org/0000-0002-7690-8360}{ORCID})
}

}
\keyword{internal}
