% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{timestamp}
\alias{timestamp}
\alias{timestamp_from}
\alias{timestamp_to}
\title{Handling Timestamps}
\usage{
timestamp_from(ms, tz = NULL, ...)

timestamp_to(dt)
}
\arguments{
\item{ms}{Unix time in milliseconds.}

\item{tz}{Time zone, when \code{NULL} it defaults to \code{sdmevaltool_options()$tz}.}

\item{...}{Other optional arguments passed to \code{\link[=as.POSIXct]{as.POSIXct()}}.}

\item{dt}{POSIXct coercible date/time value.}
}
\value{
\code{timestamp_from} returns \code{\link[=POSIXct]{POSIXct()}}.

\code{timestamp_to} returns Unix time as character .
}
\description{
Handling Timestamps
}
\examples{
dt1 <- as.POSIXct("2020-12-24 21:15:49 MDT")
dt2 <- timestamp_to(dt1)
dt3 <- timestamp_from(dt2)
stopifnot(identical(dt3, dt1))

}
\seealso{
\code{\link[=sdmevaltool_options]{sdmevaltool_options()}} to set \code{tz} globally.
}
