% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/materials.R
\name{upload-materials}
\alias{upload-materials}
\alias{prep_predictor_raster}
\alias{prep_predictor_metadata}
\alias{prep_model_metadata}
\alias{prep_spatial_prediction}
\alias{prep_observations}
\alias{prep_model_summary}
\alias{prep_model_fit}
\title{Prepare Upload Materials}
\usage{
prep_predictor_raster(
  x,
  resolution = 5,
  model_id,
  user_id,
  material_settings = "[]",
  con = NULL,
  update = FALSE,
  ...
)

prep_predictor_metadata(
  x,
  model_id,
  user_id,
  material_settings = "[]",
  con = NULL,
  update = FALSE,
  ...
)

prep_model_metadata(
  x,
  model_id,
  user_id,
  material_settings = "[]",
  con = NULL,
  update = FALSE,
  ...
)

prep_spatial_prediction(
  x,
  model_id,
  species_id,
  user_id,
  material_settings = "[]",
  con = NULL,
  update = FALSE,
  ...
)

prep_observations(
  x,
  species,
  model_id,
  species_id,
  user_id,
  material_settings = "[]",
  con = NULL,
  update = FALSE,
  ...
)

prep_model_summary(
  x,
  species,
  model_id,
  species_id,
  user_id,
  material_settings = "[]",
  con = NULL,
  update = FALSE,
  ...
)

prep_model_fit(
  x,
  species,
  model_id,
  species_id,
  user_id,
  material_settings = "[]",
  con = NULL,
  update = FALSE,
  ...
)
}
\arguments{
\item{x}{An object to upload.}

\item{resolution}{Raster resampling resolution}

\item{model_id}{Model ID.}

\item{user_id}{User ID.}

\item{material_settings}{Material settings.}

\item{con}{Database connection (use \code{NULL} to avoid inserting new record).}

\item{update}{Logical. Is this a new entry or an update to an existing one.}

\item{...}{Arguments passed to the write function.}

\item{species_id}{Species ID.}

\item{species}{Reference species table.}
}
\description{
Prepare Upload Materials
}
